fix: delete book caused by update book mechanism
The deletion process didnt work as intended. Turns out, it caused by the update step. The update step was instead inserting the updated object with IDENTICAL ID (which was forbidden) rather than replacing it. As a result, object duplicates exist. Well, the delete process was removing only one of them and left the other one. This is why if we checked the same ID again, it will return the the object